requirements.txt
yt-dlp==2023.3.4
google-api-python-client==2.86.0
python-dotenv==0.21.1